# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Persona Debate Local API
  version: 1.0.0
  description: |
    Local-first Node.js API for persona/debate/chat/governance workflows.
    All responses use a standard envelope:
    - success: { ok: true, data: ... }
    - error: { ok: false, error: { code, message, details } }
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
  - cookieAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/personas:
    get:
      tags: [Personas]
      summary: List personas
      parameters:
        - in: query
          name: q
          schema: { type: string }
        - in: query
          name: tag
          schema: { type: string }
      responses:
        '200':
          description: Personas list
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
    post:
      tags: [Personas]
      summary: Create persona
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Persona' }
            examples:
              sample:
                value:
                  id: frank-bot
                  displayName: Franck
                  role: Planner
                  description: Event-planning specialist persona
                  systemPrompt: You are Franck, practical and expressive.
                  speakingStyle:
                    tone: friendly
                    verbosity: concise
                    quirks: [theatrical emphasis]
                  expertiseTags: [event planning, aesthetics]
                  biasValues: [quality-first]
                  debateBehavior: Steelman first, then challenge assumptions.
                  knowledgePackIds: []
      responses:
        '201':
          description: Persona created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
  /api/personas/{id}:
    get:
      tags: [Personas]
      summary: Get persona
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Persona
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    put:
      tags: [Personas]
      summary: Update persona
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Persona' }
      responses:
        '200':
          description: Persona updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
        '400': { $ref: '#/components/responses/ErrorResponse' }
        '404': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      tags: [Personas]
      summary: Delete persona
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Persona deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/personas/{id}/duplicate:
    post:
      tags: [Personas]
      summary: Duplicate persona
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [id]
              properties:
                id: { type: string }
                displayName: { type: string }
      responses:
        '201':
          description: Duplicated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
  /api/personas/generate-from-topic:
    post:
      tags: [Personas]
      summary: Generate persona drafts from topic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic]
              properties:
                topic: { type: string }
                context: { type: string, default: '' }
                count: { type: integer, minimum: 1, maximum: 6, default: 3 }
                model: { type: string, default: gpt-4.1-mini }
                sources:
                  type: array
                  items: { $ref: '#/components/schemas/TopicSource' }
      responses:
        '200':
          description: Draft personas
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OkResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: pd_session
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    OkResponse:
      type: object
      properties:
        ok: { type: boolean, example: true }
        data:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      properties:
        ok: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details:
              nullable: true
              oneOf:
                - type: 'null'
                - type: array
                  items:
                    type: object
                    properties:
                      path: { type: string }
                      message: { type: string }
    SpeakingStyle:
      type: object
      properties:
        tone: { type: string, default: '' }
        verbosity: { type: string, default: '' }
        quirks:
          type: array
          items: { type: string }
      required: [tone, verbosity, quirks]
    Persona:
      type: object
      required: [id, displayName, systemPrompt, speakingStyle, expertiseTags, biasValues, debateBehavior, knowledgePackIds]
      properties:
        id:
          type: string
          pattern: '^[a-z0-9]+(?:-[a-z0-9]+)*$'
        displayName: { type: string }
        role: { type: string, default: '' }
        description: { type: string, default: '' }
        systemPrompt: { type: string }
        speakingStyle: { $ref: '#/components/schemas/SpeakingStyle' }
        expertiseTags:
          type: array
          items: { type: string }
        biasValues:
          oneOf:
            - type: array
              items: { type: string }
            - type: string
        debateBehavior: { type: string, default: '' }
        knowledgePackIds:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    TopicSource:
      type: object
      required: [title, url]
      properties:
        title: { type: string }
        source: { type: string, default: '' }
        url: { type: string, format: uri }
        publishedAt:
          oneOf:
            - type: string
            - type: 'null'
        snippet: { type: string, default: '' }
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'